checkout minor change

+++ Diamond.DataAccess/Repositories/VnPayRepository.cs
@@ -28,7 +28,7 @@ namespace DiamondShop.Controllers
 			pay.AddRequestData("vnp_Version", _configuration["Vnpay:Version"]);
 			pay.AddRequestData("vnp_Command", _configuration["Vnpay:Command"]);
 			pay.AddRequestData("vnp_TmnCode", _configuration["Vnpay:TmnCode"]);
-			pay.AddRequestData("vnp_Amount", ((int)model.price * 100).ToString());
+			pay.AddRequestData("vnp_Amount", (model.price * 100).ToString());
 			pay.AddRequestData("vnp_CreateDate", timeNow.ToString("yyyyMMddHHmmss"));
 			pay.AddRequestData("vnp_CurrCode", _configuration["Vnpay:CurrCode"]);
 			pay.AddRequestData("vnp_IpAddr", pay.GetIpAddress(context));
